requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argparse==1.4.0
asttokens==2.4.1
async-generator==1.10
atomicwrites==1.4.1
attrs==22.2.0
backcall==0.2.0
bashutils==0.0.4
beautifulsoup4==4.11.2
bleach==4.1.0
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
catboost==1.1.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.5
contextvars==2.4
cycler==0.11.0
cymem==2.0.7
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==2.0.1
fast-histogram==0.11
gym==0.26.2
gym-notices==0.0.8
highway-env==1.7.1
icecream==2.1.3
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
intel-openmp==2023.0.0
ipykernel==5.3.4
ipython==7.16.3
ipython_genutils==0.2.0
ipywidgets==7.7.5
jedi==0.17.2
jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter_client==7.1.2
jupyter_core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.4
keras==2.10.0
kiwisolver==1.3.1
layers==0.1.5
llvmlite==0.36.0
lxml==4.9.2
markupsafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl==2021.4.0
mpl-scatter-density==0.7
mpmath==1.2.1
murmurhash==1.0.9
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
openpyxl==3.1.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathy==0.10.1
patsy==0.5.5
pickleshare==0.7.5
pillow==8.4.0
pip==21.2.2
plotly==5.14.1
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.36
protobuf==3.19.6
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pydot==1.4.2
pygame==2.5.1
pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.4
pyyaml==6.0
pyzmq==22.2.1
requests==2.27.1
scikit-learn==0.24.2
scikit-posthocs==0.8.1
scipy==1.5.4
seaborn==0.11.2
send2trash==1.8.2
sentencepiece==0.1.97
setuptools==58.0.4
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.0.9
spacy-legacy==3.0.12
srsly==2.4.6
stable-baselines3==1.3.0
statannotations==0.6.0
statsmodels==0.12.2
sympy==1.9
tenacity==8.2.2
tensorboardx==2.4
terminado==0.12.1
testpath==0.6.0
thinc==8.0.17
threadpoolctl==3.1.0
tomli==1.2.3
torch==1.10.2
torchaudio==0.10.2
torchtext==0.6.0
torchvision==0.11.3
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.6.4
wincertstore==0.2
xgboost==1.5.2
xlrd==1.2.0
zipp==3.6.0
zlib==1.2.11
